Move P2P code from Vr to Main

+++ android/app/src/main/java/com/ishinomakihackathon2016/bluespring/util/Api.java
@@ -76,7 +76,7 @@ public class Api {
         post("/peer/register", body, callback);
     }
 
-    public void getPeer(int roomId, ApiCallback callback) {
+    public void getPeer(long roomId, ApiCallback callback) {
         // リクエストボディを作る
         RequestBody body = new FormEncodingBuilder().build();
